+++ tests/fixtures/Simple/SimpleEnum.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace Tests\PHPat\fixtures\Simple;
+
+enum SimpleEnum
+{
+    case Foo;
+    case Bar;
+}