+++ docs/CHANGELOG.md
@@ -1,6 +1,44 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 0.12.0
+* Remove deprecated selector modifiers
+* Bump minimum PHP version to 8.1
+* Bump minimum PHPStan version to 2.1
+* Add `ShouldBeEnum` and `ShouldNotBeEnum` assertions
+
+## 0.11.10
+* Fix `ShouldHaveOnlyOnePublicMethodNamed` assertion
+
+## 0.11.9
+* Fix `ShouldBeNamed` assertion
+* Add `isStandardClass` selector
+
+## 0.11.8
+* Add PHP 8.3 missing built-in classes
+* Catch instanceof in dependency rules
+
+## 0.11.7
+* Add `withFilepath` selector
+* Rollback param name for `ShouldBeNamed` assertion
+
+## 0.11.6
+* Add missing rule identifier to Declaration assertions
+* Fix support for regex in `ShouldHaveOnlyOnePublicMethodNamed`
+* Add support for attribute arguments in `AppliesAttribute` assertion
+
+## 0.11.5
+* Add `ShouldBeInvokable` assertion
+* Fix incorrect param name for `ShouldBeNamed` assertion
+
+## 0.11.4
+* Add `ShouldHaveOnlyPublicMethodNamed` assertion
+* Fix run on previous PHP versions
+
+## 0.11.3
+* Fix error with getNativeReflection in latest PHPStan versions
+* Catch included traits in dependency assertions
+
 ## 0.11.2
 * Add `AllOf()` selector modifier
 * Add `AnyOf()` selector modifier